+++ web/api/v1/api.go
@@ -39,11 +39,9 @@ import (
 	"github.com/prometheus/common/route"
 
 	"github.com/prometheus/prometheus/config"
-	"github.com/prometheus/prometheus/pkg/gate"
 	"github.com/prometheus/prometheus/pkg/labels"
 	"github.com/prometheus/prometheus/pkg/textparse"
 	"github.com/prometheus/prometheus/pkg/timestamp"
-	"github.com/prometheus/prometheus/prompb"
 	"github.com/prometheus/prometheus/promql"
 	"github.com/prometheus/prometheus/promql/parser"
 	"github.com/prometheus/prometheus/rules"
@@ -54,11 +52,6 @@ import (
 	"github.com/prometheus/prometheus/tsdb/index"
 	"github.com/prometheus/prometheus/util/httputil"
 	"github.com/prometheus/prometheus/util/stats"
-)
-
-const (
-	namespace = "prometheus"
-	subsystem = "api"
 )
 
 type status string
@@ -83,12 +76,6 @@ const (
 
 var (
 	LocalhostRepresentations = []string{"127.0.0.1", "localhost", "::1"}
-	remoteReadQueries        = prometheus.NewGauge(prometheus.GaugeOpts{
-		Namespace: namespace,
-		Subsystem: subsystem,
-		Name:      "remote_read_queries",
-		Help:      "The current number of remote read queries being executed or waiting.",
-	})
 )
 
 type apiError struct {
@@ -183,23 +170,21 @@ type API struct {
 	ready                 func(http.HandlerFunc) http.HandlerFunc
 	globalURLOptions      GlobalURLOptions
 
-	db                        TSDBAdminStats
-	dbDir                     string
-	enableAdmin               bool
-	logger                    log.Logger
-	remoteReadSampleLimit     int
-	remoteReadMaxBytesInFrame int
-	remoteReadGate            *gate.Gate
-	CORSOrigin                *regexp.Regexp
-	buildInfo                 *PrometheusVersion
-	runtimeInfo               func() (RuntimeInfo, error)
-	gatherer                  prometheus.Gatherer
-	remoteWriteHandler        http.Handler
+	db          TSDBAdminStats
+	dbDir       string
+	enableAdmin bool
+	logger      log.Logger
+	CORSOrigin  *regexp.Regexp
+	buildInfo   *PrometheusVersion
+	runtimeInfo func() (RuntimeInfo, error)
+	gatherer    prometheus.Gatherer
+
+	remoteWriteHandler http.Handler
+	remoteReadHandler  http.Handler
 }
 
 func init() {
 	jsoniter.RegisterTypeEncoderFunc("promql.Point", marshalPointJSON, marshalPointJSONIsEmpty)
-	prometheus.MustRegister(remoteReadQueries)
 }
 
 // NewAPI returns an initialized API type.
@@ -233,23 +218,22 @@ func NewAPI(
 		targetRetriever:       tr,
 		alertmanagerRetriever: ar,
 
-		now:                       time.Now,
-		config:                    configFunc,
-		flagsMap:                  flagsMap,
-		ready:                     readyFunc,
-		globalURLOptions:          globalURLOptions,
-		db:                        db,
-		dbDir:                     dbDir,
-		enableAdmin:               enableAdmin,
-		rulesRetriever:            rr,
-		remoteReadSampleLimit:     remoteReadSampleLimit,
-		remoteReadGate:            gate.New(remoteReadConcurrencyLimit),
-		remoteReadMaxBytesInFrame: remoteReadMaxBytesInFrame,
-		logger:                    logger,
-		CORSOrigin:                CORSOrigin,
-		runtimeInfo:               runtimeInfo,
-		buildInfo:                 buildInfo,
-		gatherer:                  gatherer,
+		now:              time.Now,
+		config:           configFunc,
+		flagsMap:         flagsMap,
+		ready:            readyFunc,
+		globalURLOptions: globalURLOptions,
+		db:               db,
+		dbDir:            dbDir,
+		enableAdmin:      enableAdmin,
+		rulesRetriever:   rr,
+		logger:           logger,
+		CORSOrigin:       CORSOrigin,
+		runtimeInfo:      runtimeInfo,
+		buildInfo:        buildInfo,
+		gatherer:         gatherer,
+
+		remoteReadHandler: remote.NewReadHandler(logger, q, configFunc, remoteReadSampleLimit, remoteReadConcurrencyLimit, remoteReadMaxBytesInFrame),
 	}
 
 	if ap != nil {
@@ -317,7 +301,7 @@ func (api *API) Register(r *route.Router) {
 	r.Get("/status/buildinfo", wrap(api.serveBuildInfo))
 	r.Get("/status/flags", wrap(api.serveFlags))
 	r.Get("/status/tsdb", wrap(api.serveTSDBStatus))
-	r.Post("/read", api.ready(http.HandlerFunc(api.remoteRead)))
+	r.Post("/read", api.ready(http.HandlerFunc(api.remoteReadHandler.ServeHTTP)))
 	r.Post("/write", api.ready(http.HandlerFunc(api.remoteWrite)))
 
 	r.Get("/alerts", wrap(api.alerts))
@@ -331,7 +315,6 @@ func (api *API) Register(r *route.Router) {
 	r.Put("/admin/tsdb/delete_series", wrap(api.deleteSeries))
 	r.Put("/admin/tsdb/clean_tombstones", wrap(api.cleanTombstones))
 	r.Put("/admin/tsdb/snapshot", wrap(api.snapshot))
-
 }
 
 type queryData struct {
@@ -1316,214 +1299,6 @@ func (api *API) serveTSDBStatus(*http.Request) apiFuncResult {
 		MemoryInBytesByLabelName:    convertStats(s.IndexPostingStats.LabelValueStats),
 		SeriesCountByLabelValuePair: convertStats(s.IndexPostingStats.LabelValuePairsStats),
 	}, nil, nil, nil}
-}
-
-func (api *API) remoteRead(w http.ResponseWriter, r *http.Request) {
-	ctx := r.Context()
-	if err := api.remoteReadGate.Start(ctx); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	remoteReadQueries.Inc()
-
-	defer api.remoteReadGate.Done()
-	defer remoteReadQueries.Dec()
-
-	req, err := remote.DecodeReadRequest(r)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusBadRequest)
-		return
-	}
-
-	externalLabels := api.config().GlobalConfig.ExternalLabels.Map()
-
-	sortedExternalLabels := make([]prompb.Label, 0, len(externalLabels))
-	for name, value := range externalLabels {
-		sortedExternalLabels = append(sortedExternalLabels, prompb.Label{
-			Name:  name,
-			Value: value,
-		})
-	}
-	sort.Slice(sortedExternalLabels, func(i, j int) bool {
-		return sortedExternalLabels[i].Name < sortedExternalLabels[j].Name
-	})
-
-	responseType, err := remote.NegotiateResponseType(req.AcceptedResponseTypes)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusBadRequest)
-		return
-	}
-
-	switch responseType {
-	case prompb.ReadRequest_STREAMED_XOR_CHUNKS:
-		api.remoteReadStreamedXORChunks(ctx, w, req, externalLabels, sortedExternalLabels)
-	default:
-		// On empty or unknown types in req.AcceptedResponseTypes we default to non streamed, raw samples response.
-		api.remoteReadSamples(ctx, w, req, externalLabels, sortedExternalLabels)
-	}
-}
-
-func (api *API) remoteReadSamples(
-	ctx context.Context,
-	w http.ResponseWriter,
-	req *prompb.ReadRequest,
-	externalLabels map[string]string,
-	sortedExternalLabels []prompb.Label,
-) {
-	w.Header().Set("Content-Type", "application/x-protobuf")
-	w.Header().Set("Content-Encoding", "snappy")
-
-	resp := prompb.ReadResponse{
-		Results: make([]*prompb.QueryResult, len(req.Queries)),
-	}
-	for i, query := range req.Queries {
-		if err := func() error {
-			filteredMatchers, err := filterExtLabelsFromMatchers(query.Matchers, externalLabels)
-			if err != nil {
-				return err
-			}
-
-			querier, err := api.Queryable.Querier(ctx, query.StartTimestampMs, query.EndTimestampMs)
-			if err != nil {
-				return err
-			}
-			defer func() {
-				if err := querier.Close(); err != nil {
-					level.Warn(api.logger).Log("msg", "Error on querier close", "err", err.Error())
-				}
-			}()
-
-			var hints *storage.SelectHints
-			if query.Hints != nil {
-				hints = &storage.SelectHints{
-					Start:    query.Hints.StartMs,
-					End:      query.Hints.EndMs,
-					Step:     query.Hints.StepMs,
-					Func:     query.Hints.Func,
-					Grouping: query.Hints.Grouping,
-					Range:    query.Hints.RangeMs,
-					By:       query.Hints.By,
-				}
-			}
-
-			var ws storage.Warnings
-			resp.Results[i], ws, err = remote.ToQueryResult(querier.Select(false, hints, filteredMatchers...), api.remoteReadSampleLimit)
-			if err != nil {
-				return err
-			}
-			for _, w := range ws {
-				level.Warn(api.logger).Log("msg", "Warnings on remote read query", "err", w.Error())
-			}
-			for _, ts := range resp.Results[i].Timeseries {
-				ts.Labels = remote.MergeLabels(ts.Labels, sortedExternalLabels)
-			}
-			return nil
-		}(); err != nil {
-			if httpErr, ok := err.(remote.HTTPError); ok {
-				http.Error(w, httpErr.Error(), httpErr.Status())
-				return
-			}
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-	}
-
-	if err := remote.EncodeReadResponse(&resp, w); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-}
-
-func (api *API) remoteReadStreamedXORChunks(ctx context.Context, w http.ResponseWriter, req *prompb.ReadRequest, externalLabels map[string]string, sortedExternalLabels []prompb.Label) {
-	w.Header().Set("Content-Type", "application/x-streamed-protobuf; proto=prometheus.ChunkedReadResponse")
-
-	f, ok := w.(http.Flusher)
-	if !ok {
-		http.Error(w, "internal http.ResponseWriter does not implement http.Flusher interface", http.StatusInternalServerError)
-		return
-	}
-
-	for i, query := range req.Queries {
-		if err := func() error {
-			filteredMatchers, err := filterExtLabelsFromMatchers(query.Matchers, externalLabels)
-			if err != nil {
-				return err
-			}
-
-			querier, err := api.Queryable.ChunkQuerier(ctx, query.StartTimestampMs, query.EndTimestampMs)
-			if err != nil {
-				return err
-			}
-			defer func() {
-				if err := querier.Close(); err != nil {
-					level.Warn(api.logger).Log("msg", "Error on chunk querier close", "err", err.Error())
-				}
-			}()
-
-			var hints *storage.SelectHints
-			if query.Hints != nil {
-				hints = &storage.SelectHints{
-					Start:    query.Hints.StartMs,
-					End:      query.Hints.EndMs,
-					Step:     query.Hints.StepMs,
-					Func:     query.Hints.Func,
-					Grouping: query.Hints.Grouping,
-					Range:    query.Hints.RangeMs,
-					By:       query.Hints.By,
-				}
-			}
-
-			ws, err := remote.StreamChunkedReadResponses(
-				remote.NewChunkedWriter(w, f),
-				int64(i),
-				// The streaming API has to provide the series sorted.
-				querier.Select(true, hints, filteredMatchers...),
-				sortedExternalLabels,
-				api.remoteReadMaxBytesInFrame,
-			)
-			if err != nil {
-				return err
-			}
-
-			for _, w := range ws {
-				level.Warn(api.logger).Log("msg", "Warnings on chunked remote read query", "warnings", w.Error())
-			}
-			return nil
-		}(); err != nil {
-			if httpErr, ok := err.(remote.HTTPError); ok {
-				http.Error(w, httpErr.Error(), httpErr.Status())
-				return
-			}
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-	}
-}
-
-// filterExtLabelsFromMatchers change equality matchers which match external labels
-// to a matcher that looks for an empty label,
-// as that label should not be present in the storage.
-func filterExtLabelsFromMatchers(pbMatchers []*prompb.LabelMatcher, externalLabels map[string]string) ([]*labels.Matcher, error) {
-	matchers, err := remote.FromLabelMatchers(pbMatchers)
-	if err != nil {
-		return nil, err
-	}
-
-	filteredMatchers := make([]*labels.Matcher, 0, len(matchers))
-	for _, m := range matchers {
-		value := externalLabels[m.Name]
-		if m.Type == labels.MatchEqual && value == m.Value {
-			matcher, err := labels.NewMatcher(labels.MatchEqual, m.Name, "")
-			if err != nil {
-				return nil, err
-			}
-			filteredMatchers = append(filteredMatchers, matcher)
-		} else {
-			filteredMatchers = append(filteredMatchers, m)
-		}
-	}
-
-	return filteredMatchers, nil
 }
 
 func (api *API) remoteWrite(w http.ResponseWriter, r *http.Request) {
